refactor(cli): Refactor CLI utils
